move to new directory, fix up the status conditions (still todos on this), handle results
