+++ control-plane/catalog/registration/registrations_controller_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package configentries_test
+package registration_test
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 	"strings"
 	"testing"
 
-	mapset "github.com/deckarep/golang-set"
+	mapset "github.com/deckarep/golang-set/v2"
 	logrtest "github.com/go-logr/logr/testing"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
@@ -30,8 +30,8 @@ import (
 	"github.com/hashicorp/go-uuid"
 
 	"github.com/hashicorp/consul-k8s/control-plane/api/v1alpha1"
+	"github.com/hashicorp/consul-k8s/control-plane/catalog/registration"
 	"github.com/hashicorp/consul-k8s/control-plane/consul"
-	"github.com/hashicorp/consul-k8s/control-plane/controllers/configentries"
 	"github.com/hashicorp/consul-k8s/control-plane/helper/test"
 )
 
@@ -65,7 +65,7 @@ func TestReconcile_Success(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:       "test-registration",
-					Finalizers: []string{configentries.RegistrationFinalizer},
+					Finalizers: []string{registration.RegistrationFinalizer},
 				},
 				Spec: v1alpha1.RegistrationSpec{
 					ID:         "node-id",
@@ -95,13 +95,21 @@ func TestReconcile_Success(tt *testing.T) {
 				},
 			},
 			serverResponseConfig: serverResponseConfig{registering: true},
-			expectedFinalizers:   []string{configentries.RegistrationFinalizer},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionTrue,
-				Reason:  "",
-				Message: "",
-			}},
+			expectedFinalizers:   []string{registration.RegistrationFinalizer},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:    v1alpha1.ConditionSynced,
+					Status:  v1.ConditionTrue,
+					Reason:  "",
+					Message: "",
+				},
+				{
+					Type:    registration.ConditionRegistered,
+					Status:  v1.ConditionTrue,
+					Reason:  "",
+					Message: "",
+				},
+			},
 		},
 		"registering -- ACLs enabled and policy does not exist": {
 			registration: &v1alpha1.Registration{
@@ -111,7 +119,7 @@ func TestReconcile_Success(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:       "test-registration",
-					Finalizers: []string{configentries.RegistrationFinalizer},
+					Finalizers: []string{registration.RegistrationFinalizer},
 				},
 				Spec: v1alpha1.RegistrationSpec{
 					ID:         "node-id",
@@ -144,13 +152,27 @@ func TestReconcile_Success(tt *testing.T) {
 				registering: true,
 				aclEnabled:  true,
 			},
-			expectedFinalizers: []string{configentries.RegistrationFinalizer},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionTrue,
-				Reason:  "",
-				Message: "",
-			}},
+			expectedFinalizers: []string{registration.RegistrationFinalizer},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:    v1alpha1.ConditionSynced,
+					Status:  v1.ConditionTrue,
+					Reason:  "",
+					Message: "",
+				},
+				{
+					Type:    registration.ConditionRegistered,
+					Status:  v1.ConditionTrue,
+					Reason:  "",
+					Message: "",
+				},
+				{
+					Type:    registration.ConditionACLsUpdated,
+					Status:  v1.ConditionTrue,
+					Reason:  "",
+					Message: "",
+				},
+			},
 		},
 		"registering -- ACLs enabled and policy does exists": {
 			registration: &v1alpha1.Registration{
@@ -160,7 +182,7 @@ func TestReconcile_Success(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:       "test-registration",
-					Finalizers: []string{configentries.RegistrationFinalizer},
+					Finalizers: []string{registration.RegistrationFinalizer},
 				},
 				Spec: v1alpha1.RegistrationSpec{
 					ID:         "node-id",
@@ -194,13 +216,27 @@ func TestReconcile_Success(tt *testing.T) {
 				aclEnabled:   true,
 				policyExists: true,
 			},
-			expectedFinalizers: []string{configentries.RegistrationFinalizer},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionTrue,
-				Reason:  "",
-				Message: "",
-			}},
+			expectedFinalizers: []string{registration.RegistrationFinalizer},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:    v1alpha1.ConditionSynced,
+					Status:  v1.ConditionTrue,
+					Reason:  "",
+					Message: "",
+				},
+				{
+					Type:    registration.ConditionRegistered,
+					Status:  v1.ConditionTrue,
+					Reason:  "",
+					Message: "",
+				},
+				{
+					Type:    registration.ConditionACLsUpdated,
+					Status:  v1.ConditionTrue,
+					Reason:  "",
+					Message: "",
+				},
+			},
 		},
 		"deregistering": {
 			registration: &v1alpha1.Registration{
@@ -210,7 +246,7 @@ func TestReconcile_Success(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:              "test-registration",
-					Finalizers:        []string{configentries.RegistrationFinalizer},
+					Finalizers:        []string{registration.RegistrationFinalizer},
 					DeletionTimestamp: &deletionTime,
 				},
 				Spec: v1alpha1.RegistrationSpec{
@@ -254,7 +290,7 @@ func TestReconcile_Success(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:              "test-registration",
-					Finalizers:        []string{configentries.RegistrationFinalizer},
+					Finalizers:        []string{registration.RegistrationFinalizer},
 					DeletionTimestamp: &deletionTime,
 				},
 				Spec: v1alpha1.RegistrationSpec{
@@ -311,12 +347,12 @@ func TestReconcile_Success(tt *testing.T) {
 				WithStatusSubresource(&v1alpha1.Registration{}).
 				Build()
 
-			controller := &configentries.RegistrationsController{
+			controller := &registration.RegistrationsController{
 				Client: fakeClient,
 				Log:    logrtest.NewTestLogger(t),
 				Scheme: s,
-				Cache: &configentries.RegistrationCache{
-					Services:            mapset.NewSet(),
+				Cache: &registration.RegistrationCache{
+					Services:            mapset.NewSet[string](),
 					ConsulClientConfig:  testClient.Cfg,
 					ConsulServerConnMgr: testClient.Watcher,
 				},
@@ -361,7 +397,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:       "test-registration",
-					Finalizers: []string{configentries.RegistrationFinalizer},
+					Finalizers: []string{registration.RegistrationFinalizer},
 				},
 				Spec: v1alpha1.RegistrationSpec{
 					ID:         "node-id",
@@ -394,12 +430,18 @@ func TestReconcile_Failure(tt *testing.T) {
 				registering:   true,
 				errOnRegister: true,
 			},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionFalse,
-				Reason:  configentries.ConsulErrorRegistration,
-				Message: "",
-			}},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:   v1alpha1.ConditionSynced,
+					Status: v1.ConditionFalse,
+					Reason: registration.SyncError,
+				},
+				{
+					Type:   registration.ConditionRegistered,
+					Status: v1.ConditionFalse,
+					Reason: registration.ConsulErrorRegistration,
+				},
+			},
 		},
 		"registering - terminating gateway acl role not found": {
 			registration: &v1alpha1.Registration{
@@ -409,7 +451,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:       "test-registration",
-					Finalizers: []string{configentries.RegistrationFinalizer},
+					Finalizers: []string{registration.RegistrationFinalizer},
 				},
 				Spec: v1alpha1.RegistrationSpec{
 					ID:         "node-id",
@@ -443,12 +485,22 @@ func TestReconcile_Failure(tt *testing.T) {
 				aclEnabled:       true,
 				temGWRoleMissing: true,
 			},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionFalse,
-				Reason:  configentries.ConsulErrorACL,
-				Message: "",
-			}},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:   v1alpha1.ConditionSynced,
+					Status: v1.ConditionFalse,
+					Reason: registration.SyncError,
+				},
+				{
+					Type:   registration.ConditionRegistered,
+					Status: v1.ConditionTrue,
+				},
+				{
+					Type:   registration.ConditionACLsUpdated,
+					Status: v1.ConditionFalse,
+					Reason: registration.ConsulErrorACL,
+				},
+			},
 		},
 		"registering - error reading policy": {
 			registration: &v1alpha1.Registration{
@@ -458,7 +510,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:       "test-registration",
-					Finalizers: []string{configentries.RegistrationFinalizer},
+					Finalizers: []string{registration.RegistrationFinalizer},
 				},
 				Spec: v1alpha1.RegistrationSpec{
 					ID:         "node-id",
@@ -493,12 +545,22 @@ func TestReconcile_Failure(tt *testing.T) {
 				errOnPolicyRead: true,
 				policyExists:    true,
 			},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionFalse,
-				Reason:  configentries.ConsulErrorACL,
-				Message: "",
-			}},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:   v1alpha1.ConditionSynced,
+					Status: v1.ConditionFalse,
+					Reason: registration.SyncError,
+				},
+				{
+					Type:   registration.ConditionRegistered,
+					Status: v1.ConditionTrue,
+				},
+				{
+					Type:   registration.ConditionACLsUpdated,
+					Status: v1.ConditionFalse,
+					Reason: registration.ConsulErrorACL,
+				},
+			},
 		},
 		"registering - policy does not exist - error creating policy": {
 			registration: &v1alpha1.Registration{
@@ -508,7 +570,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:       "test-registration",
-					Finalizers: []string{configentries.RegistrationFinalizer},
+					Finalizers: []string{registration.RegistrationFinalizer},
 				},
 				Spec: v1alpha1.RegistrationSpec{
 					ID:         "node-id",
@@ -542,12 +604,22 @@ func TestReconcile_Failure(tt *testing.T) {
 				aclEnabled:       true,
 				errOnPolicyWrite: true,
 			},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionFalse,
-				Reason:  configentries.ConsulErrorACL,
-				Message: "",
-			}},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:   v1alpha1.ConditionSynced,
+					Status: v1.ConditionFalse,
+					Reason: registration.SyncError,
+				},
+				{
+					Type:   registration.ConditionRegistered,
+					Status: v1.ConditionTrue,
+				},
+				{
+					Type:   registration.ConditionACLsUpdated,
+					Status: v1.ConditionFalse,
+					Reason: registration.ConsulErrorACL,
+				},
+			},
 		},
 		"registering - error updating role": {
 			registration: &v1alpha1.Registration{
@@ -557,7 +629,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:       "test-registration",
-					Finalizers: []string{configentries.RegistrationFinalizer},
+					Finalizers: []string{registration.RegistrationFinalizer},
 				},
 				Spec: v1alpha1.RegistrationSpec{
 					ID:         "node-id",
@@ -591,12 +663,22 @@ func TestReconcile_Failure(tt *testing.T) {
 				aclEnabled:      true,
 				errOnRoleUpdate: true,
 			},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionFalse,
-				Reason:  configentries.ConsulErrorACL,
-				Message: "",
-			}},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:   v1alpha1.ConditionSynced,
+					Status: v1.ConditionFalse,
+					Reason: registration.SyncError,
+				},
+				{
+					Type:   registration.ConditionRegistered,
+					Status: v1.ConditionTrue,
+				},
+				{
+					Type:   registration.ConditionACLsUpdated,
+					Status: v1.ConditionFalse,
+					Reason: registration.ConsulErrorACL,
+				},
+			},
 		},
 		"deregistering": {
 			registration: &v1alpha1.Registration{
@@ -606,7 +688,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:              "test-registration",
-					Finalizers:        []string{configentries.RegistrationFinalizer},
+					Finalizers:        []string{registration.RegistrationFinalizer},
 					DeletionTimestamp: &deletionTime,
 				},
 				Spec: v1alpha1.RegistrationSpec{
@@ -639,12 +721,18 @@ func TestReconcile_Failure(tt *testing.T) {
 			serverResponseConfig: serverResponseConfig{
 				errOnDeregister: true,
 			},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionFalse,
-				Reason:  configentries.ConsulErrorDeregistration,
-				Message: "",
-			}},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:   v1alpha1.ConditionSynced,
+					Status: v1.ConditionFalse,
+					Reason: registration.SyncError,
+				},
+				{
+					Type:   registration.ConditionRegistered,
+					Status: v1.ConditionFalse,
+					Reason: registration.ConsulErrorRegistration,
+				},
+			},
 		},
 		"deregistering - ACLs enabled - terminating-gateway error updating role": {
 			registration: &v1alpha1.Registration{
@@ -654,7 +742,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:              "test-registration",
-					Finalizers:        []string{configentries.RegistrationFinalizer},
+					Finalizers:        []string{registration.RegistrationFinalizer},
 					DeletionTimestamp: &deletionTime,
 				},
 				Spec: v1alpha1.RegistrationSpec{
@@ -688,12 +776,22 @@ func TestReconcile_Failure(tt *testing.T) {
 				aclEnabled:      true,
 				errOnRoleUpdate: true,
 			},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionFalse,
-				Reason:  configentries.ConsulErrorACL,
-				Message: "",
-			}},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:   v1alpha1.ConditionSynced,
+					Status: v1.ConditionFalse,
+					Reason: registration.SyncError,
+				},
+				{
+					Type:   registration.ConditionRegistered,
+					Status: v1.ConditionTrue,
+				},
+				{
+					Type:   registration.ConditionACLsUpdated,
+					Status: v1.ConditionFalse,
+					Reason: registration.ConsulErrorACL,
+				},
+			},
 		},
 		"deregistering - ACLs enabled - terminating-gateway error deleting policy": {
 			registration: &v1alpha1.Registration{
@@ -703,7 +801,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				},
 				ObjectMeta: metav1.ObjectMeta{
 					Name:              "test-registration",
-					Finalizers:        []string{configentries.RegistrationFinalizer},
+					Finalizers:        []string{registration.RegistrationFinalizer},
 					DeletionTimestamp: &deletionTime,
 				},
 				Spec: v1alpha1.RegistrationSpec{
@@ -737,12 +835,22 @@ func TestReconcile_Failure(tt *testing.T) {
 				aclEnabled:        true,
 				errOnPolicyDelete: true,
 			},
-			expectedConditions: []v1alpha1.Condition{{
-				Type:    "Synced",
-				Status:  v1.ConditionFalse,
-				Reason:  configentries.ConsulErrorACL,
-				Message: "",
-			}},
+			expectedConditions: []v1alpha1.Condition{
+				{
+					Type:   v1alpha1.ConditionSynced,
+					Status: v1.ConditionFalse,
+					Reason: registration.SyncError,
+				},
+				{
+					Type:   registration.ConditionRegistered,
+					Status: v1.ConditionTrue,
+				},
+				{
+					Type:   registration.ConditionACLsUpdated,
+					Status: v1.ConditionFalse,
+					Reason: registration.ConsulErrorACL,
+				},
+			},
 		},
 	}
 
@@ -765,12 +873,12 @@ func TestReconcile_Failure(tt *testing.T) {
 				WithStatusSubresource(&v1alpha1.Registration{}).
 				Build()
 
-			controller := &configentries.RegistrationsController{
+			controller := &registration.RegistrationsController{
 				Client: fakeClient,
 				Log:    logrtest.NewTestLogger(t),
 				Scheme: s,
-				Cache: &configentries.RegistrationCache{
-					Services:            mapset.NewSet(),
+				Cache: &registration.RegistrationCache{
+					Services:            mapset.NewSet[string](),
 					ConsulClientConfig:  testClient.Cfg,
 					ConsulServerConnMgr: testClient.Watcher,
 				},
@@ -794,7 +902,7 @@ func TestReconcile_Failure(tt *testing.T) {
 				}
 			}
 
-			require.ElementsMatch(t, fetchedReg.Finalizers, []string{configentries.RegistrationFinalizer})
+			require.ElementsMatch(t, fetchedReg.Finalizers, []string{registration.RegistrationFinalizer})
 		})
 	}
 }
